+++ lib/net.dart
@@ -6,6 +6,8 @@ library net;
 export 'src/net/email.dart';
 export 'src/net/ipv4.dart';
 export 'src/net/ipv6.dart';
+export 'src/net/mac_addr.dart';
 export 'src/net/req_email.dart';
 export 'src/net/req_ipv4.dart';
 export 'src/net/req_ipv6.dart';
+export 'src/net/req_mac_addr.dart';